add validation to record inputs, move the shared validation code to the data-client-library
