+++ apps/newsletters-ui/src/app/components/StateEditForm.tsx
@@ -1,5 +1,6 @@
 import { Paper } from '@mui/material';
 import type { z } from 'zod';
+import { getValidationWarnings } from '@newsletters-nx/newsletters-data-client';
 import type { WizardFormData } from '@newsletters-nx/state-machine';
 import type { FieldDef, FieldValue } from './SchemaForm';
 import { getModification, SchemaForm } from './SchemaForm';
@@ -10,26 +11,6 @@ interface Props {
 	formData: WizardFormData;
 	setFormData: { (newData: WizardFormData): void };
 }
-
-const getValidationWarnings = (
-	formSchema: z.ZodObject<z.ZodRawShape>,
-	formData: WizardFormData,
-): Partial<Record<string, string>> => {
-	const parseResult = formSchema.safeParse(formData);
-	const validationWarnings: Partial<Record<string, string>> = {};
-
-	if (!parseResult.success) {
-		parseResult.error.issues.forEach((issue) => {
-			const { message, path, code } = issue;
-			const key = typeof path[0] === 'string' ? path[0] : undefined;
-
-			if (key) {
-				validationWarnings[key] = message || code;
-			}
-		});
-	}
-	return validationWarnings;
-};
 
 export const StateEditForm = ({ formSchema, formData, setFormData }: Props) => {
 	const changeFormData = (value: FieldValue, field: FieldDef) => {
@@ -48,7 +29,7 @@ export const StateEditForm = ({ formSchema, formData, setFormData }: Props) => {
 			<SchemaForm
 				schema={formSchema}
 				data={formData}
-				validationWarnings={getValidationWarnings(formSchema, formData)}
+				validationWarnings={getValidationWarnings(formData, formSchema)}
 				changeValue={changeFormData}
 			/>
 		</Paper>
