+++ libs/newsletters-data-client/src/lib/zod-helpers/index.ts
@@ -1,4 +1,4 @@
-//TO DO - replace duplicates in other librarys
 export * from './getEmptySchemaData';
 export * from './recursiveUnwrap';
 export * from './schema-helpers';
+export * from './validation';
